Lint and refactor tets
