+++ .jshintignore
@@ -1,4 +1,2 @@
 node_modules/**
 public/**
-bin/**
-tests/**
